+++ brain4j-core/src/main/java/net/echo/brain4j/loss/impl/BinaryCrossEntropy.java
@@ -18,7 +18,7 @@ public class BinaryCrossEntropy implements LossFunction {
     }
 
     @Override
-    public double getDelta(double error, double derivative) {
+    public float getDelta(float error, float derivative) {
         return error;
     }
 }
