+++ brain4j-core/src/main/java/net/echo/brain4j/loss/impl/CrossEntropy.java
@@ -17,7 +17,7 @@ public class CrossEntropy implements LossFunction {
     }
 
     @Override
-    public double getDelta(double error, double derivative) {
+    public float getDelta(float error, float derivative) {
         return error;
     }
 }
